[workflows] Cancel expired runs after updating a PR
Currently, when a PR is updated and the previous CI checks
for the PR have already started, the CI will needlessly
keep executing the started tasks even though their results
are no longer relevant. Incorporate:
https://github.com/rokroskar/workflow-run-cleanup-action
to remove the expired tasks from the queue so that we get
relevant CI results faster.